Hotplug implememtiert

+++ hotplug.cpp
@@ -2,59 +2,71 @@
 #include <unistd.h>
 #include <sys/socket.h>
 #include <string.h>
-
 #include <stdio.h>
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
 #include <linux/input.h>
+#include "input.h"
 
-Hotplug::Hotplug(const char* sockName)
+Hotplug::Hotplug(std::list<class InputDevice*> *hotplugDevices)
 {
-    m_sockName = sockName;
+    m_sockName = "/tmp/test.socket";
+    m_hotplugdevices = hotplugDevices;
 }
 
 Hotplug::~Hotplug()
 {
-    close(m_fd);
+   if(m_fd)
+    close(m_fd->fd);
 }
 
-/*Setup loopback interface to trigger interrupt when OCP connected.
-Ping on loobpack interface executed from external C-program via udev rule ??*/
-int Hotplug::init(){
+int Hotplug::init(struct pollfd *fd){
 
+    if (fd) {
+        m_fd = fd;
         memset((char *)&m_programmSockaddr, 0, sizeof(m_programmSockaddr));
+
         /*Open socket on local loopback interface*/
         int hotplug_socket = socket(AF_UNIX, SOCK_DGRAM, 0);
         if(hotplug_socket < 0){
                 return -1;
             }
-        m_fd = hotplug_socket;
-
+        m_fd->fd = hotplug_socket;
         m_programmSockaddr.sun_family = AF_UNIX;
         strcpy(m_programmSockaddr.sun_path,m_sockName);
         unlink(m_sockName);
         socklen_t len = sizeof (m_programmSockaddr);
-        if (bind(m_fd, (struct sockaddr *) &m_programmSockaddr, len) < 0) {
-            qDebug("failed bind socket %d: %s", m_fd, strerror(errno));
+        if (bind(m_fd->fd, (struct sockaddr *) &m_programmSockaddr, len) < 0) {
+            qDebug("failed bind socket %d: %s", m_fd->fd, strerror(errno));
             return -1;
         }
-        //m_fd->events = POLLIN | POLLPRI;
-
-       return m_fd;
+        m_fd->events = POLLIN | POLLPRI;
+    }
+       return 0;
 }
 
+bool Hotplug::eventReceived() {
+    return m_fd ? (m_fd->revents & POLLIN) || (m_fd->revents & POLLPRI) : false;
+}
 
 /*Read data on socket*/
-bool Hotplug::readEvent(){
+void Hotplug::readEvent(){
+   if(eventReceived()){
     struct sockaddr_un peer;
     memset((char *)&peer, 0, sizeof(peer));
     size_t len = sizeof (m_programmSockaddr);
     char buf[256];
-    if(recvfrom(m_fd, buf, sizeof (buf)/sizeof (buf[0]), 0, (struct sockaddr *) &peer, &len) < 0){
-        qDebug("failed read from socket %d: %s", m_fd, strerror(errno));
+    memset(buf,0,sizeof(buf)/sizeof (buf[0]));
+    if(recvfrom(m_fd->fd, buf, sizeof (buf)/sizeof (buf[0]), 0, (struct sockaddr *) &peer, &len) < 0){
+        qDebug("failed read from socket %d: %s", m_fd->fd, strerror(errno));
     }
-    qDebug("received Device: %s",buf);
+    for (InputDevice* d : *m_hotplugdevices){
+        if(strcmp(d->m_fileName,&buf[0])==0){
+            d->init(d->m_fd);
+            qDebug("received Device: %s",buf);
+        }
+    }
+}
 
-    return strcmp(m_sockName,&buf[0])==0 ? true : false;
 }
